refactor(ccusage): split codex adapter primitives
Move Codex adapter path discovery, pricing, usage arithmetic, and shared types out of the large adapter index. This makes the adapter boundary match the other agent adapters and keeps future Codex parser work localized.

Codex detection now short-circuits when it finds the first JSONL session file instead of collecting every session path just to answer a boolean. Focused Codex adapter tests and tsgo typecheck pass.
